tkt poto tout va bien

+++ events/stringSelectMenu.js
@@ -14,14 +14,14 @@ module.exports = {
         const authorId = interaction.user.id;
         const { user, customId } = interaction;
 
+        const args = customId.split('-');
+        const command = args.shift();
+
         if(!db.doesPlayerExists(user.id).then(exists => { return exists; })) {
             return;
         }
 
-        switch(customId) {
-            case 'chooseClass':
-                rpg.setClass(interaction.values[0], interaction);
-                break;
+        switch(command) {
             case 'displayInventory':
                 inv.displayInventory(authorId, interaction);
                 break;
@@ -31,7 +31,13 @@ module.exports = {
             case 'combat_target_selector':
                 combat.receiveTargetSelector(interaction);
                 break;
-            case 'class-choice':
+            case 'classChoice':
+
+                if(args[0] != interaction.user.id) {
+                    interaction.channel.send("If you would like to start your own adventure, please use the t.init commande yourself ! " + "<@" + interaction.user.id + ">");
+                    return;
+                }
+
                 await interaction.message.delete();
 
                 db.createPlayer(interaction.user.id);
@@ -46,7 +52,7 @@ module.exports = {
                 )
                 .setThumbnail(interaction.user.displayAvatarURL());
             
-                await interaction.reply({embeds: [displayEmbed]});
+                await interaction.channel.send({embeds: [displayEmbed]});
             default:
                 return;
         }
